feat(basemodel): detect aliases in cmd paths & add multi-alias support
This should have been done from the beginning, but I forgot lol. What
this means is, that any alias defined in a command path, will make the
command work with all aliases for that command.

So for instance if you have a bills command with two aliases "bills" &
"b". The command path can be either "b view" or "bills view" and it will
work for both, regardless which one you're using to define the path.