chore: pack structs better for better memory utilization
I just recently found out that struct fields are stacked next to each
other in memory and will be padded based on the size needed for the
type. Should always be ordered from largest to smallest to reduce
padding.

This is very much a micro-optimization, but I want to be mindful.