feat(cmdmodel): do not exec command on every window size msg
I originally thought that command data needed to be processed for the
view to update, but that's incorrect. The view should recalculate itself
with cached command data when the terminal is resized.

This basically eliminates re-render performance issues when the terminal
is resized. Obviously commands should be trying to cache their data as
much as reasonable, but this specifically prevents unnecesary command
executions.