refactor(sqlite): remove bill id from bill history
The bills table is explicitly for defining what bills will be eligible
for being in a users bill history. This means they're ephemeral; at any
point the user should be able to delete a bill, but not the bills
history of use, so we must store all bill data inside the bill_history
table.

5 years from now a user could have completely different bills, but their
history remains the same.